Move activity classes to activities package
