+++ app/src/main/res/layout/activity_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ListActivity">
+    tools:context=".activities.ListActivity">
 
     <FrameLayout
         android:layout_width="match_parent"
